fix(notify): resolve preventDuplicates bug with functional state update
