+++ registry/reusables/ui/notify/index.ts
@@ -2,4 +2,13 @@ export * from "./notify-provider"
 
 export { toast } from "./notify-utils"
 
-export { Toast, statusStyles } from "./notify"
+export { Toast } from "./notify"
+
+export {
+  progressBarVariants,
+  statusStyles,
+  toastActionVariants,
+  toastPositionVariants,
+  toastTypeVariants,
+  toastVariants,
+} from "./notify-variants"
