+++ registry/reusables/ui/notify/notify.tsx
@@ -1,68 +1,16 @@
 "use client"
 
 import * as React from "react"
-import { cva } from "class-variance-authority"
 import { Loader, LoaderCircle, X } from "lucide-react"
 
 import { cn } from "@/lib/utils"
 
-import type { ToastClassNames, ToastParams } from "./notify-provider"
-
-export const statusStyles = {
-  error:
-    "dark:bg-[#24161b] bg-[#fff5f5] text-red-900 dark:text-[#ffdfdd] dark:border-red-900 border-red-200",
-  warning:
-    "dark:bg-[#1E1A1B] bg-[#fefae1] text-[#3b2212] dark:text-[#EADB90] dark:border-[#5C431B] border-[#ddcab8]",
-  success:
-    "dark:bg-[#131d1e] bg-[#e7fef6] text-[#0d311e] dark:text-[#abf9de] dark:border-[#1e5643] border-green-200",
-  info: "dark:bg-[#161831] bg-[#edf4ff] text-[#1e3a8a] dark:text-[#DCE6FF] dark:border-[#2f3873] pborder-[#bfdbfe]",
-  default:
-    "dark:bg-[#13141b] bg-white text-gray-900 dark:text-[#e4e5e9] dark:border-[#3a3c4a] border-gray-200",
-  loading: "bg-white text-gray-900 border-gray-200  ",
-} as const
-
-const toastVariants = cva(
-  "relative flex w-full flex-col gap-1 overflow-hidden rounded-lg border p-[0.75rem] shadow-lg",
-  {
-    variants: {
-      status: statusStyles,
-    },
-    defaultVariants: {
-      status: "default",
-    },
-  }
-)
-
-const progressBarVariants = cva("absolute bottom-0 left-0 h-[2px]", {
-  variants: {
-    status: {
-      error: "bg-red-600 dark:bg-[#f77a6f]",
-      warning: "bg-yellow-500 dark:bg-[#fabe20]",
-      success: "bg-green-600 dark:bg-[#12f0a5]",
-      info: "bg-blue-600 dark:bg-[#7898ff]",
-      default: "bg-gray-600 dark:bg-[#e4e5e9]",
-    },
-  },
-  defaultVariants: {
-    status: "default",
-  },
-})
-
-const toastActionVariants = cva(
-  "inline-flex items-center justify-center rounded-md px-3 py-1.5 text-sm font-medium transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
-  {
-    variants: {
-      variant: {
-        primary:
-          "bg-white text-black hover:bg-white/90 dark:text-black dark:hover:bg-white/90",
-        dismiss: "bg-transparent hover:bg-black/10 dark:hover:bg-white/10",
-      },
-    },
-    defaultVariants: {
-      variant: "primary",
-    },
-  }
-)
+import { ToastClassNames, ToastParams } from "./notify-types"
+import {
+  progressBarVariants,
+  toastActionVariants,
+  toastVariants,
+} from "./notify-variants"
 
 const ToastTitle = React.forwardRef<
   HTMLParagraphElement,
@@ -212,8 +160,6 @@ export function Toast({
   hideProgressBar,
   actions,
   paused,
-  animation = "slide",
-  position,
 }: ToastParams & { classNames?: ToastClassNames; paused?: boolean }) {
   return (
     <div
